requirements.txt
asgiref==3.7.2
Django==5.0.3
dnspython==2.6.1
image==1.5.33
numpy==1.26.4
pillow==10.2.0
pymongo==4.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
